Fix double client init
