+++ apps/xmtp.chat/src/hooks/useConversations.ts
@@ -3,10 +3,10 @@ import type {
   SafeListConversationsOptions,
 } from "@xmtp/browser-sdk";
 import { useState } from "react";
-import { useClient } from "./useClient";
+import { useXMTP } from "@/contexts/XMTPContext";
 
 export const useConversations = () => {
-  const { client } = useClient();
+  const { client } = useXMTP();
   const [loading, setLoading] = useState(false);
   const [syncing, setSyncing] = useState(false);
